+++ Assets/Ocean/Scripts/Boyant.cs
@@ -8,44 +8,25 @@ public class Boyant : MonoBehaviour {
 private Ocean ocean;
 
 public float buoyancy ; //use this var to adjust the level the object floats, 0 is neutral
-public bool reactsToChoppiness ;
-public float chopinessfactor = 1f;
-private Vector3 basePosition;
-private Rigidbody rrigidbody;
-public bool useRigidbody;
+private bool hasChoppy;
+private Vector3 oldPos;
 
 
 	void Start () {
-		//used for choppy offset
-		basePosition = transform.position;
-		rrigidbody =  GetComponent<Rigidbody>();
-		if(rrigidbody == null) useRigidbody = false;
-   
 		ocean = Ocean.Singleton;
+		if(ocean.choppy_scale>0) hasChoppy = true;
+		oldPos = transform.position;
 	}
 	
 	void Update () {
 		if(ocean.canCheckBuoyancyNow) {
+			float off = 0;
+			if(hasChoppy) off = ocean.GetChoppyAtLocation(transform.position.x, transform.position.z);
 
-			if(useRigidbody) {
-
-				rrigidbody.AddForceAtPosition (-Vector3.left * (ocean.GetChoppyAtLocation(basePosition.x, basePosition.z) * chopinessfactor*Random.Range(0.1f,0.5f)), transform.position);
-				float targetY = ocean.GetWaterHeightAtLocation2(transform.position.x, transform.position.z) + buoyancy;
-				transform.position = new Vector3(transform.position.x , targetY, basePosition.z);
-
-			} else {
-
-				if(!reactsToChoppiness){ basePosition = transform.position; }
-				float targetX = basePosition.x;
-
-				if(reactsToChoppiness){
-					targetX += ocean.GetChoppyAtLocation2(basePosition.x, basePosition.z) * chopinessfactor;
-				}
-
-				float targetY = ocean.GetWaterHeightAtLocation2(targetX, basePosition.z) + buoyancy;
-				transform.position = new Vector3(targetX , targetY, basePosition.z);
-			}
-		}
+			float targetY = ocean.GetWaterHeightAtLocation2(transform.position.x-off, transform.position.z) + buoyancy;
+			transform.position = new Vector3(transform.position.x, targetY, transform.position.z);
+			oldPos = transform.position;
+		} else {transform.position = oldPos;}
 	}